Integrate logout into homepage
